refactor(registrations): lift questions domain to src/lib per /simplify
Move EVENT_REGISTRATION_QUESTIONS + RegistrationQuestions +
RegistrationAnswer + validateRegistrationQuestions + buildRegistrationAnswers
out of Events/eventOptions.ts into src/lib/registrations/questions.ts, so
Registrations no longer cross-imports from the Events collection (the
project-structure rule's cross-owner-import violation). RegistrationAnswer
is now defined once (the email imports it from src/lib). Also simplify
buildRegistrationAnswers to a two-loop form (drops the seen-set, the push
helper, and the QUESTION_LABELS map).

Refs #588
