+++ src/collections/Events/eventOptions.ts
@@ -44,105 +44,3 @@ export const EVENT_REGISTRATION_MODE_OPTIONS = [
   { label: 'Sahaj Atlas', value: 'sahaj-atlas' },
   { label: 'External', value: 'external' },
 ] as const
-
-/**
- * Optional questions an event can ask registrants. Rendered as a group of
- * checkboxes on the Event (each label IS the question shown to the registrant);
- * checking one includes that question on the registration form. Placeholder
- * wording — refine per the real registration flow.
- */
-export const EVENT_REGISTRATION_QUESTIONS = [
-  { name: 'priorExperience', label: 'Have you practised Sahaja Yoga meditation before?' },
-  { name: 'referralSource', label: 'How did you hear about this event?' },
-  { name: 'healthInfo', label: 'Is there anything about your health we should know?' },
-  { name: 'accessibility', label: 'Do you have any accessibility requirements?' },
-  { name: 'guests', label: 'Will you be bringing any guests?' },
-] as const
-
-/** A configured registration-question key (`priorExperience`, `referralSource`, …). */
-export type EventRegistrationQuestionName = (typeof EVENT_REGISTRATION_QUESTIONS)[number]['name']
-
-/**
- * The stored shape of a registration's `questions` field: each configured
- * question key maps to the registrant's (string) answer. Every key is optional —
- * an event enables only some questions, and a registrant may skip any.
- */
-export type RegistrationQuestions = Partial<Record<EventRegistrationQuestionName, string>>
-
-const EVENT_REGISTRATION_QUESTION_NAMES = new Set<string>(
-  EVENT_REGISTRATION_QUESTIONS.map((question) => question.name),
-)
-
-/**
- * Enforce the `questions` structure on write: an object whose keys are all
- * configured question names and whose values are strings. Returns `true` or an
- * error message. Pure JS (no `jsonSchema` — Payload would compile that to an Ajv
- * `new Function()` validator, which throws under Cloudflare's codegen ban; see
- * `src/fields/translationsField.ts`).
- */
-export function validateRegistrationQuestions(value: unknown): true | string {
-  if (value == null) return true
-  if (typeof value !== 'object' || Array.isArray(value)) {
-    return 'Registration answers must be an object of question answers.'
-  }
-  for (const [key, answer] of Object.entries(value)) {
-    if (!EVENT_REGISTRATION_QUESTION_NAMES.has(key)) {
-      return `Unknown registration question: "${key}".`
-    }
-    if (typeof answer !== 'string') {
-      return `The answer for "${key}" must be text.`
-    }
-  }
-  return true
-}
-
-/** A registrant's answer to one registration question, labelled for display. */
-export interface RegistrationAnswer {
-  label: string
-  value: string
-}
-
-const QUESTION_LABELS = new Map<string, string>(
-  EVENT_REGISTRATION_QUESTIONS.map((question) => [question.name, question.label]),
-)
-
-/** Render one raw answer value as a display string, or `''` to skip it. */
-function formatAnswer(raw: unknown): string {
-  if (typeof raw === 'string') return raw.trim()
-  if (typeof raw === 'boolean') return raw ? 'Yes' : 'No'
-  if (typeof raw === 'number') return String(raw)
-  if (Array.isArray(raw))
-    return raw
-      .map((entry) => String(entry).trim())
-      .filter(Boolean)
-      .join(', ')
-  return ''
-}
-
-/**
- * Shape a registrant's raw `questions` answers (the record stored on the
- * registration) into labelled rows for the manager notification. Keys map to
- * their configured question label, falling back to the key for anything not in
- * `EVENT_REGISTRATION_QUESTIONS`; blank answers are dropped. Ordered to follow
- * the configured question order, then any extra keys.
- */
-export function buildRegistrationAnswers(
-  questions: Record<string, unknown> | null | undefined,
-): RegistrationAnswer[] {
-  if (!questions || typeof questions !== 'object') return []
-
-  const answers: RegistrationAnswer[] = []
-  const seen = new Set<string>()
-
-  const push = (key: string) => {
-    if (seen.has(key) || !(key in questions)) return
-    seen.add(key)
-    const value = formatAnswer(questions[key])
-    if (value) answers.push({ label: QUESTION_LABELS.get(key) ?? key, value })
-  }
-
-  for (const question of EVENT_REGISTRATION_QUESTIONS) push(question.name)
-  for (const key of Object.keys(questions)) push(key)
-
-  return answers
-}
